+++ CHANGELOG.md
@@ -4,6 +4,13 @@ Changelog for the Vitreus runtime.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [212] - 2025-03-05
+
+### Added
+
+- Add support for three energy assets
+- Add runtime API to retrieve supported swap paths
+
 ## [211] - 2025-02-15
 
 ### Added